uma sub pages
Due to the granularity of Windows allocation in 64 KB, the size of the free list element becomes 2 bytes. The optimization approach is as follows: split the free list and the slab elements themselves into subgroups, for which one byte is sufficient to identify a free element of the group, and also makes it possible to selectively indicate that a page is no longer needed.